move import-equals tests to their own file
runtime-transpiler.test.ts has two pre-existing tests (// @bun > async
transpiler / require()) that fail on release builds for unrelated reasons;
keep the new coverage in a dedicated file so it is independent of those.
